admin-role

+++ backend/controllers/productController.js
@@ -21,16 +21,21 @@ exports.createProduct = async (req, res, next) => {
   let id = 1;
   if (length > 0) id = products[length - 1].id + 1;
   const { name, category, new_price, old_price } = req.body;
+  const { _id } = req.user;
   try {
-    const product = await Product.create({
-      id,
-      name,
-      image: image_filename,
-      category,
-      new_price,
-      old_price,
-    });
-    res.status(200).json(product);
+    if (user && user.rol === "admin") {
+      const product = await Product.create({
+        id,
+        name,
+        image: image_filename,
+        category,
+        new_price,
+        old_price,
+      });
+      res.status(200).json(product);
+    } else {
+      return next(errorHandler(400, "You are not an admin."));
+    }
   } catch (err) {
     next(err);
   }
@@ -38,12 +43,17 @@ exports.createProduct = async (req, res, next) => {
 
 exports.deleteProduct = async (req, res, next) => {
   const { id } = req.params;
+  const { _id } = req.user;
   try {
-    const product = await Product.findOne({ id });
-    //delete the image from folder
-    fs.unlink(`upload/images/${product.image}`, () => {});
-    await Product.findOneAndDelete({ id });
-    res.status(200).json(product);
+    if (user && user.rol === "admin") {
+      const product = await Product.findOne({ id });
+      //delete the image from folder
+      fs.unlink(`upload/images/${product.image}`, () => {});
+      await Product.findOneAndDelete({ id });
+      res.status(200).json(product);
+    } else {
+      return next(errorHandler(400, "You are not an admin."));
+    }
   } catch (err) {
     next(err);
   }
